floats to decimals

+++ db/migrate/20120620185129_add_price_to_materials.rb
@@ -1,5 +1,5 @@
 class AddPriceToMaterials < ActiveRecord::Migration
   def change
-    add_column :materials, :price, :float, default: 0
+    add_column :materials, :price, :decimal, default: 0
   end
 end
