+++ db/migrate/20120620190346_add_price_to_recipes.rb
@@ -1,5 +1,5 @@
 class AddPriceToRecipes < ActiveRecord::Migration
   def change
-    add_column :recipes, :price, :float, default: 0
+    add_column :recipes, :price, :decimal, default: 0
   end
 end
